Create HTML  Git Code
This code is straight from the code that I had copied and pasted from the Adobe Software Dreamwork's. The code is not completely ours we used a template to get started quicker and then from their we will edit and modify the code in the way we want our website to look.